+++ src/EndPoints/Artifacts.php
@@ -6,7 +6,7 @@ class Artifacts extends Endpoint
 {
     const RESOURCE = 'artifacts';
 
-    public function photo()
+    public function photos()
     {
         return $this->getResource('photo');
     }
@@ -14,5 +14,15 @@ class Artifacts extends Endpoint
     public function portfolio()
     {
         return $this->getResource('portfolio');
+    }
+
+    public function editPhoto($id, $attributes)
+    {
+        return $this->putResource($id, $attributes);
+    }
+
+    public function deletePhoto($id)
+    {
+        $this->deleteResource($id);
     }
 }